Package projects moved to /src folder; .gitmodules changed for sentry-dotnet project
