+++ src/Directory.Build.props
@@ -1,0 +1,22 @@
+<Project>
+  <PropertyGroup>
+    <LangVersion>9</LangVersion>
+    <Nullable>enable</Nullable>
+    <UnityVersion>2019.4.18f1</UnityVersion>
+    <WarningsAsErrors>true</WarningsAsErrors>
+    <GenerateDependencyFile>false</GenerateDependencyFile>
+    <PackageRuntimePath>../../package/Runtime</PackageRuntimePath>
+    <PackageEditorPath>../../package/Editor</PackageEditorPath>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Roslynator.Analyzers" Version="3.1.0" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.CodeAnalysis.BannedApiAnalyzers" Version="3.3.0" PrivateAssets="All" />
+    <PackageReference Include="Nullable" Version="1.3.0" PrivateAssets="All" />
+  </ItemGroup>
+  
+  <!-- TODO: configure for MacOS, Linux. Currently for Windows only -->
+  <PropertyGroup Condition="Exists('C:\Program Files\Unity\Hub\Editor\$(UnityVersion)\Editor\Data\Managed\UnityEngine.dll')">
+    <UnityPath>C:\Program Files\Unity\Hub\Editor\$(UnityVersion)\Editor\Data\Managed</UnityPath>
+  </PropertyGroup>
+</Project>
