+++ src/Sentry.Unity.Editor/SentryCliOptions.cs
