+++ src/Sentry.Unity.Editor/SentryLogoDark.png
