+++ src/Sentry.Unity.Editor/SentryLogoLight.png
