+++ src/Sentry.Unity.Editor/SentryPreprocessBuild.cs
