+++ src/Sentry.Unity.sln
@@ -1,0 +1,70 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30523.141
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.Unity", "Sentry.Unity\Sentry.Unity.csproj", "{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.Unity.Editor", "Sentry.Unity.Editor\Sentry.Unity.Editor.csproj", "{583BBF7E-054D-414D-AAA8-47C400AB3914}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry", "sentry-dotnet\src\Sentry\Sentry.csproj", "{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9DBA6473-0B44-4301-AEF5-6B5E7A0D0E52}"
+ProjectSection(SolutionItems) = preProject
+	Directory.Build.props = Directory.Build.props
+EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x64.Build.0 = Debug|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x86.Build.0 = Debug|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x64.ActiveCfg = Release|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x64.Build.0 = Release|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x86.ActiveCfg = Release|Any CPU
+		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x86.Build.0 = Release|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x64.Build.0 = Debug|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x86.Build.0 = Debug|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|Any CPU.Build.0 = Release|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x64.ActiveCfg = Release|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x64.Build.0 = Release|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x86.ActiveCfg = Release|Any CPU
+		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x86.Build.0 = Release|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Debug|x64.Build.0 = Debug|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Debug|x86.Build.0 = Debug|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Release|x64.ActiveCfg = Release|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Release|x64.Build.0 = Release|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Release|x86.ActiveCfg = Release|Any CPU
+		{EB0E00F0-7FB8-4C77-A85E-74BC5744856F}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {12BCF000-7A19-4A05-93E2-F6E8E2AB0C4A}
+	EndGlobalSection
+EndGlobal
