+++ src/Sentry.Unity/SentryBehavior.cs
@@ -1,0 +1,18 @@
+﻿using Sentry;
+using UnityEngine;
+using Sentry.Unity;
+using System;
+using Sentry.Protocol;
+
+public class SentryBehavior : MonoBehaviour
+{
+    public void Disable()
+    {
+  
+    }
+
+    // TODO: Flush events. See note on OnApplicationQuit
+    //private void OnApplicationPause() => 
+    // TODO: Flush events, see note on OnApplicationQuit
+    // private void OnApplicationFocus { if (!focusStatus) Flush events! }
+}
