+++ src/Sentry.Unity/UnitySentryOptions.cs
