--- src/sentry-dotnet-unity/Directory.Build.props
@@ -1,34 +1,0 @@
-<Project>
-
-  <PropertyGroup>
-    <LangVersion>9</LangVersion>
-    <Nullable>enable</Nullable>
-    <UnityVersion>2019.4.11f1</UnityVersion>
-    <WarningsAsErrors>true</WarningsAsErrors>
-    <GenerateDependencyFile>false</GenerateDependencyFile>
-    <UnityPackageAssetsDirectory>../../../sentry-unity/Assets</UnityPackageAssetsDirectory>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Roslynator.Analyzers" Version="3.0.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.CodeAnalysis.BannedApiAnalyzers" Version="3.3.0" PrivateAssets="All" />
-    <PackageReference Include="Nullable" Version="1.3.0" PrivateAssets="All" />
-  </ItemGroup>
-
-  <PropertyGroup>
-    <UnityPath Condition="$(UnityPath) == '' AND Exists('C:\Program Files\Unity\Hub\Editor\$(UnityVersion)\Editor\Data\Managed\UnityEngine.dll')">C:\Program Files\Unity\Hub\Editor\$(UnityVersion)\Editor\Data\Managed</UnityPath>
-    <UnityPath Condition="$(UnityPath) == '' AND Exists('/Applications/Unity/Hub/Editor/$(UnityVersion)/Unity.app/Contents/Managed/UnityEngine.dll')">/Applications/Unity/Hub/Editor/$(UnityVersion)/Unity.app/Contents/Managed/</UnityPath>
-    <!--If not using Unity Hub, tries to pick whatever Unity version is installed on the machine-->
-    <UnityPath Condition="$(UnityPath) == '' AND Exists('C:\Program Files\Unity\Editor\Data\Managed\UnityEngine.dll')">C:\Program Files\Unity\Editor\Data\Managed</UnityPath>
-    <UnityPath Condition="$(UnityPath) == '' AND Exists('/Applications/Unity/Unity.app/Contents/Managed/UnityEngine.dll')">/Applications/Unity/Unity.app/Contents/Managed</UnityPath>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="UnityEngine">
-      <HintPath>$(UnityPath)\UnityEngine.dll</HintPath>
-      <Private>false</Private>
-    </Reference>
-  </ItemGroup>
-
-</Project>
