--- src/sentry-dotnet-unity/Sentry.Unity.Editor/Sentry.Unity.Editor.csproj
@@ -1,31 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <!-- This assembly only stays in the Editor, so embed pdbs in the dll to help debugging -->
-    <DebugType>embedded</DebugType>
-    <OutDir>$(UnityPackageAssetsDirectory)/Editor/Sentry</OutDir>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="UnityEditor">
-      <HintPath>$(UnityPath)\UnityEditor.dll</HintPath>
-      <Private>false</Private>
-    </Reference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\sentry-dotnet\src\Sentry\Sentry.csproj">
-      <Private>false</Private>
-    </ProjectReference>
-    <ProjectReference Include="..\Sentry.Unity\Sentry.Unity.csproj">
-      <Private>false</Private>
-    </ProjectReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="SentryLogoLight.png" />
-    <EmbeddedResource Include="SentryLogoDark.png" />
-  </ItemGroup>
-
-</Project>
