--- src/sentry-dotnet-unity/Sentry.Unity.sln
@@ -1,65 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30523.141
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry", "sentry-dotnet\src\Sentry\Sentry.csproj", "{A1406740-5D00-4272-9B22-FA919A10EC5C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.Unity", "Sentry.Unity\Sentry.Unity.csproj", "{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sentry.Unity.Editor", "Sentry.Unity.Editor\Sentry.Unity.Editor.csproj", "{583BBF7E-054D-414D-AAA8-47C400AB3914}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Debug|x64.Build.0 = Debug|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Debug|x86.Build.0 = Debug|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Release|x64.ActiveCfg = Release|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Release|x64.Build.0 = Release|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Release|x86.ActiveCfg = Release|Any CPU
-		{A1406740-5D00-4272-9B22-FA919A10EC5C}.Release|x86.Build.0 = Release|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x64.Build.0 = Debug|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Debug|x86.Build.0 = Debug|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x64.ActiveCfg = Release|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x64.Build.0 = Release|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x86.ActiveCfg = Release|Any CPU
-		{1325A3DF-F3C9-4EF4-94F7-B283FE9DA18D}.Release|x86.Build.0 = Release|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x64.Build.0 = Debug|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Debug|x86.Build.0 = Debug|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|Any CPU.Build.0 = Release|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x64.ActiveCfg = Release|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x64.Build.0 = Release|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x86.ActiveCfg = Release|Any CPU
-		{583BBF7E-054D-414D-AAA8-47C400AB3914}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {12BCF000-7A19-4A05-93E2-F6E8E2AB0C4A}
-	EndGlobalSection
-EndGlobal
