--- src/sentry-dotnet-unity/Sentry.Unity/link.xml
@@ -1,62 +1,0 @@
-﻿<linker>
-    <assembly fullname="Sentry" preserve="all"/>
-
-    <assembly fullname="System">
-        <type fullname="System.ComponentModel.TypeConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.ArrayConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.BaseNumberConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.BooleanConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.ByteConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.CharConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.CollectionConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.ComponentConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.CultureInfoConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.DateTimeConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.DecimalConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.DoubleConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.EnumConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.ExpandableObjectConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.Int16Converter" preserve="all"/>
-        <type fullname="System.ComponentModel.Int32Converter" preserve="all"/>
-        <type fullname="System.ComponentModel.Int64Converter" preserve="all"/>
-        <type fullname="System.ComponentModel.NullableConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.SByteConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.SingleConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.StringConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.TimeSpanConverter" preserve="all"/>
-        <type fullname="System.ComponentModel.UInt16Converter" preserve="all"/>
-        <type fullname="System.ComponentModel.UInt32Converter" preserve="all"/>
-        <type fullname="System.ComponentModel.UInt64Converter" preserve="all"/>
-    </assembly>
-
-    <!--https://docs.microsoft.com/en-us/dotnet/api/system.linq.expressions.lambdaexpression.name?view=netframework-4.7.2&viewFallbackFrom=netframework-2.0-->
-    <assembly fullname="System.Core">
-        <type fullname="System.Linq.Expressions" preserve="all"/>
-        <type fullname="System.Linq.Expressions.*" preserve="all"/>
-        <type fullname="System.Linq.Expressions.Interpreter.LightLambda" preserve="all"/>
-        <type fullname="System.Linq.Expressions.LambdaExpression" preserve="all"/>
-    </assembly>
-    <assembly fullname="System.Linq.Expressions">
-        <type fullname="System.Linq.Expressions" preserve="all"/>
-        <type fullname="System.Linq.Expressions.*" preserve="all"/>
-        <type fullname="System.Linq.Expressions.Interpreter.LightLambda" preserve="all"/>
-        <type fullname="System.Linq.Expressions.LambdaExpression" preserve="all"/>
-    </assembly>
-    <assembly fullname="System.Runtime">
-        <type fullname="System.Exception" preserve="all"/>
-        <type fullname="System.Exception" preserve="all"/>
-        <type fullname="System.Diagnostics.StackTrace" preserve="all"/>
-    </assembly>
-    <assembly fullname="mscorlib">
-        <type fullname="System.Exception" preserve="all"/>
-        <type fullname="System.Diagnostics.StackTrace" preserve="all"/>
-    </assembly>
-    <assembly fullname="netstandard">
-        <type fullname="System.Linq.Expressions" preserve="all"/>
-        <type fullname="System.Linq.Expressions.*" preserve="all"/>
-        <type fullname="System.Linq.Expressions.Interpreter.LightLambda" preserve="all"/>
-        <type fullname="System.Linq.Expressions.LambdaExpression" preserve="all"/>
-        <type fullname="System.Exception" preserve="all"/>
-        <type fullname="System.Diagnostics.StackTrace" preserve="all"/>
-    </assembly>
-</linker>
